+++ src/services/handler_interface.ts
@@ -8,18 +8,20 @@ import TypeScriptHandler from "./ts_handler";
 import RustHandler from "./rust_handler";
 
 export interface ILanguageHandler {
-  buildUserPrompt(baseUserPrompt: string): Promise<{userPrompt: string, configFileUri?: Uri, config: string}>;
+  readonly language: string;
+  readonly configOptions: string[];
+
   addMissingDependencies(configFileUri: Uri, config: string, output: string): void;
 }
 
 export function getLanguageHandler(langExtName: string): ILanguageHandler {
   switch (langExtName) {
-    case "dart": return new DartHandler();
-    case "java": return new JavaHandler();
-    case "js": case "jsx": return new JavaScriptHandler();
-    case "ts": case "tsx": return new TypeScriptHandler();
-    case "py": return new PythonHandler();
-    case "rs": return new RustHandler();
-    default: return new DefaultHandler();
+    case ".dart": return new DartHandler();
+    case ".java": return new JavaHandler();
+    case ".js": case ".jsx": return new JavaScriptHandler();
+    case ".ts": case ".tsx": return new TypeScriptHandler();
+    case ".py": return new PythonHandler();
+    case ".rs": return new RustHandler();
+    default: return new DefaultHandler(`"${langExtName}"`, "NA");
   }
 }